requirements.txt
Faker==0.8.4
pytest==3.3.1
pytz==2017.2
selenium==3.8.0
Appium_Python_Client==0.25
